dbeaver/pro#10289 Core module refactoring
