+++ plugins/org.jkiss.dbeaver.core/src/org/jkiss/dbeaver/ui/services/ApplicationPolicyService.java
@@ -1,6 +1,6 @@
 /*
  * DBeaver - Universal Database Manager
- * Copyright (C) 2010-2025 DBeaver Corp and others
+ * Copyright (C) 2010-2026 DBeaver Corp and others
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jkiss.dbeaver.core.ui.services;
+package org.jkiss.dbeaver.ui.services;
 
 import org.eclipse.core.commands.Command;
 import org.eclipse.ui.commands.ICommandService;
