+++ plugins/org.jkiss.dbeaver.ui.app.standalone/src/org/jkiss/dbeaver/ui/app/standalone/update/WorkbenchInitializerUpdateCheck.java
@@ -18,11 +18,11 @@ package org.jkiss.dbeaver.ui.app.standalone.update;
 
 import org.eclipse.ui.application.IWorkbenchWindowConfigurer;
 import org.jkiss.code.NotNull;
-import org.jkiss.dbeaver.core.ui.services.ApplicationPolicyService;
-import org.jkiss.dbeaver.core.ui.services.UIServiceApplicationVersionUpdater;
 import org.jkiss.dbeaver.model.app.DBPApplication;
 import org.jkiss.dbeaver.runtime.DBWorkbench;
 import org.jkiss.dbeaver.ui.IWorkbenchWindowInitializer;
+import org.jkiss.dbeaver.ui.services.ApplicationPolicyService;
+import org.jkiss.dbeaver.ui.services.UIServiceApplicationVersionUpdater;
 
 public class WorkbenchInitializerUpdateCheck implements IWorkbenchWindowInitializer {
     @Override
